% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpdbEnrichmentAnalysis.R
\name{cpdbEnrichmentAnalysis}
\alias{cpdbEnrichmentAnalysis}
\title{CPDB Enrichment Analysis}
\usage{
cpdbEnrichmentAnalysis(
  entityType,
  accNumbers,
  mv1,
  mv2 = NULL,
  accType,
  fsetType,
  pThreshold = 1
)
}
\arguments{
\item{entityType, }{should be either 'genes' or 'metabolites'.}

\item{accNumbers}{is a list of interesting accNumbers (e.g. differentially expressed)}

\item{mv1}{is a list of mesured value 1 or fold change if m2 is null}

\item{mv2}{is a list of mesured value 2}

\item{accType}{is a valid accession number type.}

\item{fsetType, }{the type of the functional sets to be tested (as obtained with the getCpdbAvailableFsetTypes function; e.g. 'P' for pathways).}

\item{pThreshold, }{a p-value threshold, only sets with significant enrichment below or equal to this threshold will be provided.}
}
\value{
cpdbIds, mapped accession numbers
}
\description{
Performs Wilcoxon enrichment analysis of functional sets with provided physical entities from CPDB.
}
\examples{

library(PathData)

accType <- "uniprot"
accNumbers <- c("MDHM_HUMAN", "MDHC_HUMAN", "DLDH_HUMAN", "DHSA_HUMAN", "DHSB_HUMAN", "C560_HUMAN", "DHSD_HUMAN", "ODO2_HUMAN", "ODO1_HUMAN", "CISY_HUMAN", "ACON_HUMAN", "IDH3A_HUMAN", "IDH3B_HUMAN", "IDH3G_HUMAN", "SUCA_HUMAN", "SUCB1_HUMAN", "FUMH_HUMAN", "OGDHL_HUMAN", "ACOC_HUMAN", "DHTK1_HUMAN", "AMAC1_HUMAN")

mv1 <- runif(length(accNumbers))
mv2 <- runif(length(accNumbers))

eanalysis <- cpdbEnrichmentAnalysis("genes", accNumbers, mv1, mv2, "uniprot", "C", 1)
}
