% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCpdbAvailableFsetTypes.R
\name{getAvailableFsetTypes}
\alias{getAvailableFsetTypes}
\title{Get available Functional Set Types}
\usage{
getAvailableFsetTypes(entityType)
}
\arguments{
\item{entityType, }{string with entity typme, should be 'genes' or metabolites}
}
\value{
Dataframe with Functional Set types ID and descriptions associated to genes or metabolites
}
\description{
Provides a list of available functional set types such as pathways, GO categories, NESTs, ...
Parameter 'entityType' should be either 'genes' or 'metabolites'.
}
\examples{

library(PathData)

getAvailableFsetTypes("genes")
}
