% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCpdbAccessionTypes.R
\name{getCpdbAccessionTypes}
\alias{getCpdbAccessionTypes}
\title{Get CPDB Available Accession Types}
\usage{
getCpdbAccessionTypes(entityType)
}
\arguments{
\item{entityType, }{string with entity typme, should be 'genes' or metabolites}
}
\value{
array with accession types associated to genes or metabolites
}
\description{
Provides a list of different types of accession numbers (e.g. 'uniprot') that are mappable to CPDB entity IDs.
Parameter 'entityType' should be either 'genes' or 'metabolites'.
}
\examples{

library(PathData)

getCpdbAccessionTypes(entityType = "genes")
}
