% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCpdbDefaultBackgroundSize.R
\name{getCpdbDefaultBackgroundSize}
\alias{getCpdbDefaultBackgroundSize}
\title{Get default Background Size}
\usage{
getCpdbDefaultBackgroundSize(fsetType, accType)
}
\arguments{
\item{fsetType, }{string with functional Set types ID associated to genes or metabolites}

\item{accType, }{string with accession type}
}
\value{
numeric with background size for over-representation analysis
}
\description{
Provides the default background size for over-representation analysis. Note that it is accession number-specific.
}
\examples{

library(PathData)

getCpdbDefaultBackgroundSize("P", "hgnc-symbol")
}
