% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCpdbVersion.R
\name{getCpdbVersion}
\alias{getCpdbVersion}
\title{Get CpdbVersion}
\usage{
getCpdbVersion()
}
\value{
string with ConsensusPathDB version
}
\description{
Provides the version of ConsensusPathDB which is the source for functional sets.
}
\examples{

library(PathData)

getCpdbVersion()
}
